ch(refactor test config)
