+++ package.json
@@ -4,8 +4,8 @@
   "description": "Store Manager is a web application that allows store owners to manage product sales, inventory and \r records",
   "main": "app.js",
   "scripts": {
-    "test": "export current_env=test || SET \"current_env=test\" && npm run build && nyc --reporter=lcov --reporter=text --cache mocha dist/test/*.test.js --exit --timeout=20000",
-    "test-dev": "export current_env=test || SET \"current_env=test\" && npm run test-tables-setup && npm run build && nyc --reporter=lcov --reporter=text --cache mocha dist/test/*.test.js --exit --timeout=20000",
+    "test-dev": "export current_env=test || SET \"current_env=test\" && npm run build && nyc --reporter=lcov --reporter=text --cache mocha dist/test/*.test.js --exit --timeout=20000",
+    "test": "export current_env=test || SET \"current_env=test\" && npm run test-tables-setup && npm run build && nyc --reporter=lcov --reporter=text --cache mocha dist/test/*.test.js --exit --timeout=20000",
     "start": "npm run build && node dist/app.js",
     "lint": "./node_modules/.bin/eslint server --fix",
     "clean": "rm -rf dist",
